fix Controllers
