+++ src/main/java/com/moskalenko/bankcinema/api/client/ActorClient.java
@@ -1,0 +1,16 @@
+package com.moskalenko.bankcinema.api.client;
+
+import com.moskalenko.bankcinema.api.DTO.ActorDTO;
+import com.moskalenko.bankcinema.api.entity.Actor;
+
+import java.util.Collection;
+
+public interface ActorClient {
+    Actor addActor(ActorDTO actorData);
+
+    void addActorToMovie(Long actorId, Long movieId);
+
+    Actor getActorById(Long actorId);
+
+    Collection<Actor> getAllActors();
+}
