+++ src/main/java/com/moskalenko/bankcinema/api/client/DirectorClient.java
@@ -1,0 +1,14 @@
+package com.moskalenko.bankcinema.api.client;
+
+import com.moskalenko.bankcinema.api.DTO.DirectorDTO;
+import com.moskalenko.bankcinema.api.entity.Director;
+
+import java.util.Collection;
+
+public interface DirectorClient {
+    Director addDirector(DirectorDTO directorData);
+
+    Director getDirectorById(Long directorId);
+
+    Collection<Director> getAllDirectors();
+}
