+++ src/main/java/com/moskalenko/bankcinema/service/ActorService.java
@@ -2,6 +2,7 @@ package com.moskalenko.bankcinema.service;
 
 import com.moskalenko.bankcinema.api.DTO.ActorDTO;
 import com.moskalenko.bankcinema.api.entity.Actor;
+import com.moskalenko.bankcinema.api.entity.Movie;
 import com.moskalenko.bankcinema.dao.ActorDAO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -16,9 +17,11 @@ public class ActorService {
 
     private static final Logger log = LoggerFactory.getLogger(ActorService.class);
     private final ActorDAO actorDAO;
+    private final MovieService movieService;
 
-    public ActorService(ActorDAO actorDAO) {
+    public ActorService(ActorDAO actorDAO, MovieService movieService) {
         this.actorDAO = actorDAO;
+        this.movieService = movieService;
     }
 
     @Transactional
@@ -41,10 +44,22 @@ public class ActorService {
 
     public Actor getActorById(Long actorId) {
         final Actor actor = actorDAO.getActorById(actorId).orElse(null);
-        if (actor == null){
+        if (actor == null) {
             log.info("[{}] Actor is not found", actorId);
             throw new RuntimeException("Actor is not found");
         }
         return actor;
+    }
+
+    public void addActorToMovie(Long actorId, Long movieId) {
+        final Actor actor = actorDAO.getActorById(actorId).orElse(null);
+        if (actor == null) {
+            log.info("[{}] Actor is not found", actorId);
+            throw new RuntimeException("Actor is not found");
+        }
+        final Movie movie = movieService.getMovieById(movieId);
+        actor.getMovies().add(movie);
+        log.info("[{}] Movie added to [{}] actor", movieId, actorId);
+        actorDAO.save(actor);
     }
 }
